requirements.txt
pytest==7.4.2
pyyaml